requirements.txt
suds==0.4
mysql-connector==2.1.4
mysql-python==1.2.5
tqdm==4.9.0
python-dateutil==2.6.0
stompest==2.2.6
pika==0.10.0
celery==4.0.2
Django==1.10.6
djangorestframework==3.6.2
django-oauth-toolkit==1.11.0
django-cors-middleware==1.3.1
markdown==2.6.8
django-filter==1.0.1
pillow==4.0.0
